Move all source files to a src/core subdirectory; new top-level cmakelists
