+++ src/core/BinaryReader.h
