+++ src/core/BlockState.h
