+++ src/core/SqObject.cpp
