+++ src/core/SqObject.h
